refactor(pds/__init__.py and create_raw.py): Consolidated some fucntions to the top-level module.
